session_timeout should be uint64_t as uint32_t can overflow.

+++ src/rist-private.h
@@ -571,7 +571,7 @@ struct rist_peer {
 	/* Timers */
 	uint32_t rtcp_keepalive_interval;
 	uint64_t keepalive_next_time;
-	uint32_t session_timeout;
+	uint64_t session_timeout;
 	uint64_t last_rtcp_received;
 	uint64_t last_sender_report_time;
 	uint64_t last_sender_report_ts;
